Inclusão de Regras.

+++ source/Hermes/src/br/estacio/hermes/util/Util.java
@@ -4,7 +4,7 @@ import java.util.Calendar;
 
 public class Util {
 
-	public static long diffOfDays(Calendar date1, Calendar date2) {
+	public static int diffOfDays(Calendar date1, Calendar date2) {
 		// Get date in milliseconds
 		long milisecond1 = date1.getTimeInMillis();
 		long milisecond2 = date2.getTimeInMillis();
@@ -14,7 +14,7 @@ public class Util {
 
 		// Find date difference in days
 		// (24 hours 60 minutes 60 seconds 1000 millisecond)
-		long diffOfDays = (int)(diffInMSec / (24 * 60 * 60 * 1000));
+		int diffOfDays = (int)(diffInMSec / (24 * 60 * 60 * 1000));
 
 		return diffOfDays;
